+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.  
 This project adheres to [Semantic Versioning](https://semver.org).
 
+## [3.1.0] - 2023-03-29
+
+### Added
+
+- Command line option: --overwrite-directory
+- Writing current.cif and current.mtz while running
+
 ## [3.0.0] - 2023-03-09
 
 ### Added